fixes missing suspense for character create, adds build to quality checks gh action
